Crud notification added

+++ app/src/main/java/com/example/sl3verbeterd/ui/applicant/ApplicantsActivity.kt
@@ -2,10 +2,24 @@ package com.example.sl3verbeterd.ui.applicant
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AlertDialog
 import com.example.sl3verbeterd.R
+import android.app.PendingIntent
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.example.sl3verbeterd.HireHubApplication
+import com.example.sl3verbeterd.Profile
+import com.example.sl3verbeterd.ProfileListAdapter
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 
 import android.app.Activity
-import android.content.Intent
 import android.app.Application
 import android.text.TextUtils
 import android.widget.Button
@@ -14,11 +28,7 @@ import android.widget.Toast
 import androidx.activity.viewModels
 
 import androidx.lifecycle.observe
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.sl3verbeterd.HireHubApplication
-import com.example.sl3verbeterd.Profile
-import com.example.sl3verbeterd.ProfileListAdapter
+
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 
 /**
@@ -66,8 +76,21 @@ class ApplicantsActivity : AppCompatActivity(), ProfileListAdapter.ProfileClickL
     }
 
     override fun onDeleteClick(profile: Profile) {
-        applicantsViewModel.deleteProfile(profile)
+        // Create a confirmation dialog
+        AlertDialog.Builder(this)
+            .setTitle("Delete Profile")
+            .setMessage("Are you sure you want to delete ${profile.firstName} ${profile.lastName}?")
+            .setPositiveButton("Yes") { _, _ ->
+                // User confirmed, delete the profile here
+                applicantsViewModel.deleteProfile(profile)
+            }
+            .setNegativeButton("No") { dialog, _ ->
+                // User cancelled the delete operation, dismiss the dialog
+                dialog.dismiss()
+            }
+            .show()
     }
+
